Xamarin
This includes the Xamarin Form cross-platform files. There's an additional Android set of files "ghAPI.Android" that was too large for me to upload. I didn't do any changes in there so I don't think it's necessary for me to include it anyway. But let me know if you need to see the Android files.